aicc-python-api/requirements.txt
linkedin-api @ git+https://github.com/tomquirk/linkedin-api.git@766f56199b69b9c6ed988080bec19d77ad9bc03d
fastapi[all]
PyMuPDF==1.23.4
Pillow==9.5.0
rich==13.6.0
PyYAML==6.0.1
pytesseract==0.3.10